requirements.txt
graphql-core
pandas
matplotlib
